ext4_utils: add BoardConfig define to suppress EMMC-corrupting wipe command
If BOARD_SUPPRESS_EMMC_WIPE is true, the EMMC wipe command will not be
issued. This works around a bug in some firmware revisions of Samsung EMMC's
that permanently damages the device when the wipe command is issued.

For affected devices with kernel source, MMC_CAP_ERASE should be removed
from the kernel instead.

This is only part of the solution but it does handle the "flashing CM9
for the first time on an unsafe kernel" situation.

Change-Id: Ie4e31f9268a65218e5d344ae3068b021790fc33c
Signed-off-by: bgos10 <bishalgoswami@gmail.com>